Update meta-imx downstream repository branch to latest release
Use the "kirkstone-5.15.71-2.2.0" branch of meta-imx for downstream
BSP OE builds, as this is slightly more up to date.

Add another set of workarounds for broken meta-imx layer, this time
OpenCV recipe depends on non-existent recipe in another layer, which
has since moved to newer version, use the newer version and ignore
the one in meta-imx. The deepview-rt blob deployment recipe is also
broken and completely fails on package QA and later on image install,
disable it for now.

Signed-off-by: Marek Vasut <marex@denx.de>