Drop meta-dhsom layers branch refspecs
The meta-dhsom layers all have branches named 'dunfell' and 'kirkstone'
which contain identical commits and can be built for both OE 'dunfell'
and OE 'kirkstone' releases. The legacy 'dunfell-3.1' branch contains
identical commits too, but is slowly being decommissioned. The commit
SHA1 of the tip of each branch is identical to the other branches.

Use the branch matching selected OE release to avoid confusion in the
bitbake output.

Signed-off-by: Marek Vasut <marex@denx.de>