Make meta-python2 compatible with OE Scarthgap
This is a temporary adjustment to make meta-python2 work with OE 5.0 Scarthgap
to permit build of meta-qt5 with all of its components. Note that both python2
and meta-qt5 are being phased out, and receive little to no maintenance. Do not
use these for any new development. Qt5 has long been superseded by Qt6, so was
meta-qt5 by meta-qt6, which is supported by this repository.

Signed-off-by: Marek Vasut <marex@denx.de>