README: Add usage of pipx for installation on Debian 12
Starting with Debian 12 the python environment is marked as externaly
managed. Thus installing kas with pip3 fails with the error message:

error: externally-managed-environment

× This environment is externally managed
╰─> To install Python packages system-wide, try apt install
    python3-xyz, where xyz is the package you are trying to
    install.

    If you wish to install a non-Debian-packaged Python package,
    create a virtual environment using python3 -m venv path/to/venv.
    Then use path/to/venv/bin/python and path/to/venv/bin/pip. Make
    sure you have python3-full installed.

    If you wish to install a non-Debian packaged Python application,
    it may be easiest to use pipx install xyz, which will manage a
    virtual environment for you. Make sure you have pipx installed.

    See /usr/share/doc/python3.11/README.venv for more information.

note: If you believe this is a mistake, please contact your Python
installation or OS distribution provider. You can override this, at
the risk of breaking your Python installation or OS, by passing
--break-system-packages.
hint: See PEP 668 for the detailed specification.

For this case pipx exists to install python packages into local bin but
isolates it and its dependencies into a venv. In Debian 11 it has to be
installed from backports.

So update the README to use pipx for installation on newer Debian
systems.

Signed-off-by: Pascal Zimmermann <pzimmermann@dh-electronics.com>