Generate list of current state of CVEs
Generate list of current state of patched and unpatched CVEs in
the currently built system image. This is useful for analysis of
any potential security vulnerabilities, as well as tracking the
current state of CVEs applicable to this particular image build.

The list is available in deploy images/ directory as <image-name>*.json
file as well as deploy cve/ directory for each package.

Signed-off-by: Marek Vasut <marex@nabladev.com>